src/plugins/characterai/__init__.py
from pathlib import Path

import nonebot
from nonebot import get_plugin_config
from nonebot.plugin import PluginMetadata

from .config import Config

__plugin_meta__ = PluginMetadata(
    name="CharacterAI",
    description="使用CharacterAI对话",
    usage="@Bot 你的问题",
    config=Config,
)


config = get_plugin_config(Config)

sub_plugins = nonebot.load_plugins(
    str(Path(__file__).parent.joinpath("plugins").resolve())
)

