pyproject.toml
[project]
name = "XineYc_v7"
version = "0.1.0"
description = "XineYc_v7"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "characterai>=1.0.0",
    "nonebot-adapter-qq>=1.5.3",
    "nonebot2[aiohttp,fastapi]>=2.3.3",
    "psutil>=6.1.0",
    "py-cpuinfo>=9.0.0",
    "pymongo>=4.10.1",
    "requests>=2.32.3",
]

[tool.nonebot]
adapters = [
    { name = "QQ", module_name = "nonebot.adapters.qq" }
]
plugins = []
plugin_dirs = ["src/plugins"]
builtin_plugins = []
